+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "editor.codeActionsOnSave": {
+        "source.organizeImports": "explicit"
+    }
+}
