+++ package.json
@@ -16,6 +16,7 @@
     "bootstrap": "^5.3.2",
     "cdbreact": "^1.5.18",
     "cors": "^2.8.5",
+    "dotenv": "^16.4.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.9.1",
     "react-confirm-alert": "^3.0.6",
