--- src/.env
@@ -1,1 +1,0 @@
-// bring the URLs here
