--- src/components/Course/CourseCard.js
@@ -1,7 +1,0 @@
-const CourseCard = ({ course, onEdit, onDelete, onCreateNew }) => {
-    return (
-        console.log("course card", ...course)
-
-    );
-}
-export default CourseCard;
