--- src/config.js
@@ -1,7 +1,0 @@
-const config = {
-    API_URL: 'https://localhost/api',
-    // PROFS_API_URL: 'https://141.64.64.103/api',
-
-};
-
-export default config;
