+++ src/index.js
@@ -2,10 +2,10 @@ import React from 'react';
 // import { configureStore } from '@reduxjs/toolkit';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-import { store } from './redux/store'
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
+import App from './App';
+import { store } from './react/store';
+import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
