--- src/models/userModel.ts
@@ -1,5 +1,0 @@
-export interface UserModel {
-  userId: string;
-  firstName: string;
-  lastName: string;
-}
