+++ src/react/degreeCourse/redux/slices/degreeCoursesSlice.js
