+++ src/react/user/components/DeleteUserConfirmationDialog.js
